requirements.txt
psycopg2-binary==2.9.11
pymodbus==3.12.1
pyserial==3.5
python-dotenv==1.2.2
